Add files via upload
This is a sample project for browsing Images on your filesystem. Filesystem is shown as a treeview on the left, Image files and directories in selected directory is shown as a list view. 


On doubleclick image file listview will show preview of image. You can also preview any image by dropping it by dragging from treeview, listview or external location using mouse.

ImageExplorer.app.zip is an executable for this project